refactor: improve just file
