+++ nobodywho/Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "core",
     "godot",
+    "unity",
 ]
 
 # edition = "2021" implies resolver = "2".
